♻️ refactor: checkInputValue로 함수 분리 지난 로또 번호와 로또 구입 금액 입력값을 확인 할때에 공통적으로 number와 양수인지 확인하는 작업이 필요하여 checkInputValue로 분리
